Add messageTypeId and messageTsUs proprties only if they exist.
