+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.2.0",
   "dependencies": {
     "@alertlogic/al-azure-collector-js": "^1.1.2",
+    "@alertlogic/al-collector-js": "^1.1.2",
     "async": "^2.6.1",
     "moment": "^2.24.0",
     "parse-key-value": "^1.0.0"
